refactor(api): move checkCombinedCoursesFeatureIsEnabled to quest context
